@@ -4,12 +4,12 @@ apply plugin: 'maven'
 def aarName = "pinlibrary"
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '29.0.3'
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode project.ext.versionCode
         versionName project.ext.versionName
         version = android.defaultConfig.versionName
@@ -34,9 +34,9 @@ android {
 }
 
 dependencies {
-    implementation "com.android.support:support-v4:27.1.1"
-    implementation "com.android.support:appcompat-v7:27.1.1"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     androidTestImplementation 'com.jayway.android.robotium:robotium-solo:5.2.1'
 }
 
