@@ -1,7 +1,7 @@
 package com.venmo.android.pin;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 
 import com.venmo.android.pin.test.R;
 
